feat(bindforge): split BindForgeTabPanel into BindingProfile + BindingManagement sub-tabs
Extracts the binding editor from the monolithic BindForgeTabPanel into two focused
panels under a new elite.intel.ui.screen.bindforge package:

- BindingProfilePanel: the full binding editor (directory/profile/file fields,
  Used/Missing Bindings tables, conflict detection with hover callouts, conflicts-only
  filter, Fix Missing/Revert/Apply footer). All V1.1 binding-editor enhancements
  (conflict-map upgrade, hover callouts, renderBindingTables() extraction, reserved key
  chords, ship/SRV twin recommendations) are ported from the former BindForgeTabPanel
  into this panel.
- BindingManagementPanel: placeholder shell for the forthcoming backup/restore feature.
- BindForgeTabPanel: reduced to a 56-line sub-tab shell delegating to the two panels.
