+++ app/src/main/resources/i18n/gui.properties
@@ -104,6 +104,9 @@ player.conversationMode=Conversation Mode
 player.companionMode=Companion Mode
 
 # Bindings tab
+bindForge.tab.bindingProfile=Binding Profile
+bindForge.tab.bindingManagement=Binding Management
+bindForge.bindingManagement.comingSoon=Coming soon
 bindings.section.profile=Binding Profile
 bindings.section.diagnostics=Binding Diagnostics
 bindings.profileName=Profile
