+++ app/src/main/resources/i18n/gui_de.properties
@@ -54,6 +54,8 @@ player.conversationMode=Konversationssmodus
 player.companionMode=Begleitermodus
 
 # Bindings tab
+bindForge.tab.bindingProfile=Bindungsprofil
+bindForge.tab.bindingManagement=Bindungsverwaltung
 bindings.section.profile=Bindungsprofil
 bindings.section.diagnostics=Bindungsdiagnose
 bindings.profileName=Profil
