+++ app/src/main/resources/i18n/gui_es.properties
@@ -41,6 +41,8 @@ player.commandLanguage=Idioma
 player.commandLanguage.tooltip=Idioma usado por los comandos de voz y la interfaz
 player.conversationMode=Modo de conversación
 player.companionMode=Modo de compañero
+bindForge.tab.bindingProfile=Perfil de Atajos
+bindForge.tab.bindingManagement=Gestión de Atajos
 bindings.profileName=Perfil
 bindings.filePath=Archivo
 bindings.info.title=Información del perfil de bindings
