+++ app/src/main/resources/i18n/gui_fr.properties
@@ -132,6 +132,8 @@ player.companionMode=Mode compagnon
 # 06 - RACCOURCIS ELITE DANGEROUS
 # Détection et diagnostic des commandes configurées dans le jeu.
 # =============================================================================
+bindForge.tab.bindingProfile=Profil de Raccourcis
+bindForge.tab.bindingManagement=Gestion des Raccourcis
 bindings.section.profile=Profil de raccourcis
 bindings.section.diagnostics=Diagnostic des raccourcis
 bindings.profileName=Profil
