+++ app/src/main/resources/i18n/gui_pt.properties
@@ -92,6 +92,8 @@ player.commandLanguage.tooltip=Language used by voice commands and the GUI
 player.conversationMode=Conversation Mode
 player.companionMode=Companion Mode
 # Bindings tab
+bindForge.tab.bindingProfile=Perfil de Atalhos
+bindForge.tab.bindingManagement=Gerenciamento de Atalhos
 bindings.section.profile=Binding Profile
 bindings.section.diagnostics=Binding Diagnostics
 bindings.profileName=Profile
