+++ app/src/main/resources/i18n/gui_ru.properties
@@ -54,6 +54,8 @@ player.conversationMode=Режим разговора
 player.companionMode=Режим компаньона
 
 # Bindings tab
+bindForge.tab.bindingProfile=Профиль привязок
+bindForge.tab.bindingManagement=Управление привязками
 bindings.section.profile=Профиль привязок
 bindings.section.diagnostics=Диагностика привязок
 bindings.profileName=Профиль
