+++ app/src/main/resources/i18n/gui_uk.properties
@@ -54,6 +54,8 @@ player.conversationMode=Режим розмови
 player.companionMode=Режим компаньйона
 
 # Bindings tab
+bindForge.tab.bindingProfile=Профіль прив'язок
+bindForge.tab.bindingManagement=Управління прив'язками
 bindings.section.profile=Профіль прив'язок
 bindings.section.diagnostics=Діагностика прив'язок
 bindings.profileName=Профіль
